Programa final de Agenda con cambios menores

+++ agenda/src/com/generation/ejercicio/informacion/Contacto.java
@@ -31,7 +31,7 @@ public class Contacto {
         return telefono;
     }
 
-    public void setTelefono(int telefono) {
+    public void setTelefono(long telefono) {
         this.telefono = telefono;
     }
 }
